pyproject.toml
[project]
name = "instagram-sync"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "django==5.2.1",
    "djangorestframework==3.16.0",
    "psycopg2-binary==2.9.10",
    "requests==2.32.3",
    "django-environ==0.12.0",
    "drf-spectacular==0.28.0",
    "gunicorn==21.2.0",
    "whitenoise==6.6.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit==4.2.0",
    "ruff==0.11.6",
    "mypy==1.15.0",
    "types-requests==2.32.0.20250515",
    "django-stubs==5.2.0",
    "djangorestframework-stubs==3.16.0",
    "pytest==8.3.5",
    "pytest-django==4.10.0",
    "requests-mock==1.12.1",
]

[tool.mypy]
python_version = "3.11"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
strict = true
ignore_missing_imports = true
exclude = ["migrations", "tests"]

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "core.settings"
